02/pyproject.toml
[project]
name = "project02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "masiff346@gmail.com", email = "76251742+m-asif1@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.9",
]

[project.scripts]
project02 = "project02:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
